# documentation/OpenApi3.yaml
openapi: 3.0.0
info:
  title: SOFTENG24-48
  version: '1.0'
servers:
- url: http://localhost:9115
  variables: {}
paths:
...
  /api/tollStationPasses/{tollStationID}/{date_from}/{date_to}:
    get:
      tags:
      - Passes
      summary: tollStationPasses
      operationId: getTollStationPasses
      parameters:
        - name: tollStationID
          in: path
          required: true
          description: The ID of the toll station
          schema:
            type: string
        - name: date_from
          in: path
          required: true
          description: Start date in YYYYMMDD format
          schema:
            type: string
            format: date
            pattern: '^\d{8}$'
            example: '20220101'
        - name: date_to
          in: path
          required: true
          description: End date in YYYYMMDD format
          schema:
            type: string
            format: date
            pattern: '^\d{8}$'
            example: '20220131'
      responses:
        '200':
          description: 'Successfully retrieved toll station passes'
          content:
            application/json:
              schema:
                type: object
                properties:
                  StationID:
                    type: string
                  stationOperator:
                    type: string
                  requestTimestamp:
                    type: string
                    format: date-time
                  periodFrom:
                    type: string
                    format: date
                  periodTo:
                    type: string
                    format: date
                  nPasses:
                    type: integer
                  passList:
                    type: array
                    items:
                      type: object
                      properties:
                        passIndex:
                          type: integer
                        passID:
                          type: integer
                        timestamp:
                          type: string
                          format: date-time
                        tagID:
                          type: string
                        tagProvider:
                          type: string
                        passType:
                          type: string
                          enum: ["home", "visitor"]
                        passCharge:
                          type: number
                          format: float
        '204':
          description: 'No passes found'
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "failed"
                  message:
                    type: string
                    example: "No passes found for given station and time period"
        '400':
          description: 'Bad request - Invalid parameters'
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "failed"
                  message:
                    type: string
                    example: "Missing required parameters or invalid date format"
        '500':
          description: 'Internal server error'
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "failed"
                  message:
                    type: string
                    example: "Internal server error"
      deprecated: false
...